Add initial code and first test
This change adds code to download, unpack and run a single Tomcat
instance (currently 7.0.50) with a test war file deployed, and to
execute a single, simple POST and GET test against the test application.

It includes the code necessary to build the test application, which
stores and returns session data. The test application will be built on
each run if its source has changed or the war file does not exist.

The project is configured to use Rubocop and Travis CI.