Add first tests
Previously the tests included in the system test were placeholders to
make sure the infrastructure was working. This change removes those
tests and replaces them with proper tests of functionality:

 - Configure host

 - Configure password

 - Configure port

 - Configure timeout

 - Configure uri

 - Configure redis database

All but the last of these are negative tests; they show that if you
configure the wrong port, for example, Tomcat won't start. They are done
this way because the cost of creating a configurable redis environment
in Travis outweighs the added test benefit.

This change adds a redis helper (to configure the database) and a
session helper (to handle cookies and related data). It also adds to the
Tomcat helper to handle tests that should cause Tomcat to fail