Switch to using context.xml for configurations
Previously the context information was added to the server.xml file.
While supported, this is not the preferred method of defining context
in Tomcat. This change moves to using a separate context.xml file and
a simple server.xml file.

It also moves to the likely initial classname for the session manager.