Download and use the Redis Manager jar
This change enables the configure host test.

To support the tests it downloads the Redis Manager jar from S3 and
copies it to the lib directory. Currently it does this with hard-coded
version information.

It temporarily removes Tomcat 8 from the coverage matrix, pending
changes to support it.

It updates the default database number from 7 to the new value of 0, the
format of the uri config, and some expected error messages.

The change also refactors the Rakefile for readability, and consolidates
test file preparation prior to starting Tomcat into one method.

[#64604666]