Try different invalid host name
Previously the test for an invalid host used 127.1.2.3. This fails as
expected when run locally, but on Travis it connects successfully (i.e.
not the expected behaviour). It is possible that this is because of a
host mapping within Travis, so this change substitutes an obviously
incorrect host name ("foo") instead.

[#64604666]