Add test for invalid URI
This change enables the test for an invalid uri. It uses an invalid
password as this gives a more distinctive error message than the other
parameters that could be misconfigured, so the context.xml and expected
error messages are also updated.

It also fixes a problem in the Rakefile that could lead to the incorrect
Redis Manager jar being used, and updates to the latest Redis Manager.

[#64604702]