Automate the retrieval of the latest Redis Store
Previously the version of the Redis Store to be retrieved was hard-coded
in the Rakefile. This made choosing a version relatively simple, but was
error-prone. This change adds code to interrogate s3 for the latest
build, with the option to override with a specified version. It uses
similar logic to retrieve the Tomcat tar, though the version number
still needs to be specified in code (this should change infrequently).

It also:

 - rolls back to RSpec 3.0.0.beta1, as beta2 causes errors when running
in IntelliJ

 - modifies the sample app to exclude the previously embedded Tomcat,
allowing it to run against Tomcat 8, and hence adds that back to the
test matrix

 - simplifies the Rakefile by creating two rake task files.

99% of the credit for the code in this update goes to Ben Hale.