Expose application endpoints.
Previously, the test application did not expose the application's
endpoint to callers.  This was likely due to a bug in Spring Boot and
its auto-reconfiguration of embedded servlet containers (which this
application does not use).  To solve the issue and yet still support
running the test application inside of Tomcat 8, auto-reconfiguration
of embedded servlet containers was re-enabled, but the embedded Tomcat
JARs were excluded from the WAR file.  This means that the application
will no longer run stand-alone, but will run in any servlet container.

[#66846864]