Add test for Redis connection drop
This change adds a test to check that session data is not lost when the
Redis connection goes down. It does this by killing the client
connection for the session (to simulate Redis failing), and testing that
the session data is still returned and that a Redis failure is logged
in catalina.out.

The redis-rb gem used does not implement the client list or kill
commands, so these are implemented here; this code could be simplified
if redis-rb is updated with the necessary changes.

[#64605700]