Safe Tomcat download location
Previously the download location for Tomcat (from a mirrored location)
was unsafe.  When new versions of Tomcat were released, they old links
expired and builds would break.  This change updates the location of
the downloads to be from a permanent location.

This change also updates the Travis configuration to work against the
latest Tomcat 7 and 8.