Allow for different versions of Redis
Previously the test to show that Redis did not stop sessions being
returned if it went down looked for specific text to show that Redis
had failed as expected. This text has changed between versions of Jedis,
making the test brittle. This change looks for less specific text that
should still confirm the fail, making the test more robust.